analytic/requirements.txt
Flask
mysql-connector
pymongo
